update playerのUIの処理の変更 カメラの視点変更
