+++ Player.h
@@ -44,12 +44,6 @@ HRESULT InitPlayer(void);
 void UninitPlayer(void);
 void UpdatePlayer(void);
 void DrawPlayer(void);
-void DrawAttach(void);
-void DrawCapture(void);
-void DrawCollect(void);
-void DrawReload(void);
-void DrawPlayerUI(void);
-void DrawPlayerRestBullet(void);
 PLAYER* GetPlayer(void);
 XMFLOAT3 GetLerp(XMFLOAT3 initialPos, XMFLOAT3 endingPos, float percentage);
 float GetDistance3D(XMFLOAT3 pos1, XMFLOAT3 pos2);
